update ui purchasing
added alert into crud

+++ src/modules/purchasing/unit-payment-order/view.js
@@ -79,9 +79,15 @@ export class View {
     }
 
     delete(event) {
-        Promise.all([this.service.delete(this.data), this.azureService.delete(this.data)])
-            .then(result => {
-                this.cancel();
-            });
+        this.dialog.prompt('Apakah anda yakin akan menghapus data ini?', 'Hapus Data SPB')
+        .then(response => {
+            if (response.ok) {
+                Promise.all([this.service.delete(this.data), this.azureService.delete(this.data)])
+                    .then(result => {
+                        alert("Data berhasil dihapus");
+                        this.cancel();
+                    });
+            }
+        })
     }
 }
